Cache complete MatchDay responses and simplify single ZIP search
The new MatchDayService is now cache-aware and caches complete MatchDay
lists in the new 'matchDayResponse'-cache. If a 7 or 3 days search is
cached, a one day search will use a sublist of the previous results.

Single ZIP search is now handled via RegionType 'ZIP' and follows the
same approach as other region searches.